+++ tests/testthat/test-prepare_df_plot.R
@@ -34,7 +34,7 @@ test_that("TB burden data is correctly processed ready for plotting", {
                df_all$metric_label)
   expect_equal("tbl_df", class(df_all$df)[1])
   expect_equal(3651, nrow(df_all$df))
-  expect_equal(72, ncol(df_all$df))
+  expect_equal(73, ncol(df_all$df))
   expect_equal(0, sum(is.na(df_all$df$country)))
 })
 
@@ -50,7 +50,7 @@ test_that("TB burden data is correctly proccessed when using a single country", 
                df_exact_country$metric_label)
   expect_equal("tbl_df", class(df_exact_country$df)[1])
   expect_equal(17, nrow(df_exact_country$df))
-  expect_equal(72, ncol(df_exact_country$df))
+  expect_equal(73, ncol(df_exact_country$df))
   expect_equal("United Kingdom of Great Britain and Northern Ireland", 
                as.character(unique(df_exact_country$df$country)))
 })
@@ -67,7 +67,7 @@ test_that("TB burden data is correctly proccessed when comparing to the region",
                df_region$metric_label)
   expect_equal("tbl_df", class(df_region$df)[1])
   expect_equal(913, nrow(df_region$df))
-  expect_equal(72, ncol(df_region$df))
+  expect_equal(73, ncol(df_region$df))
   expect_equal(eur_countries, 
                as.character(unique(df_region$df$country))[1:10])
 })
@@ -92,6 +92,7 @@ test_df <- tibble::tibble(country = "test", year = 2000:2001,
 result_df <- test_df
 result_df$Year <- result_df$year
 result_df$country <- factor(result_df$country)
+result_df$`Estimated incidence (all forms) per 100 000 population` <- result_df$e_inc_100k
 
 
 test_that("prepare_df_plot correctly formats the input data.", {
@@ -107,9 +108,11 @@ test_that("prepare_df_plot correctly formats the input data.", {
 test_that("annual_change correctly transforms metric and confidence intervals", {
 
   result_df <- result_df[-1, ]
+  result_df <- result_df[, -ncol(result_df)]
   result_df$e_inc_100k <- 0.2
   result_df$e_inc_100k_lo <- 0.25
   result_df$e_inc_100k_hi <- 0.1
+  result_df$`Percentage annual change: Estimated incidence (all forms) per 100 000 population` <- 0.2
   
   df_annual_change <- prepare_df_plot(metric = "e_inc_100k", conf = c("_lo", "_hi"),
                                       df = test_df, annual_change = TRUE)